changed the navigation bar component to a SolidJS component
