+++ package.json
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@astrojs/mdx": "^5.0.4",
-    "@astrojs/vue": "^6.0.1",
+    "@astrojs/solid-js": "^7.0.1",
     "@mdx-js/mdx": "^3.1.1",
     "@tailwindcss/typography": "^0.5.16",
     "@tailwindcss/vite": "^4.1.13",
@@ -24,8 +24,8 @@
     "prettier-plugin-vue": "^1.1.6",
     "remark-parse": "^11.0.0",
     "sharp": "^0.34.5",
-    "tailwindcss": "^4.1.13",
-    "vue": "^3.5.33"
+    "solid-js": "^1.9.14",
+    "tailwindcss": "^4.1.13"
   },
   "devDependencies": {
     "@babel/parser": "^7.28.0",
